Eigen/g++5 conflict resolved by switching to doubles
For reasons that aren't completely clear, when optimization is turned on and
g++ version >= 5, EigenSolver returns nans if the matrices contain long
doubles.  This bit of insanity does not come up if double are used.  Since that
is preferable to either avoiding optimization or being wedded to an old
compiler, that's the option we're going with for now.

+++ AbcUtil.h
@@ -55,7 +55,7 @@ namespace ABC {
       using namespace mpfr;
       typedef mpreal float_type;
   #else
-      typedef long double float_type;
+      typedef double float_type;
   #endif 
 
   typedef Matrix<float_type,Dynamic,Dynamic> Mat2D;
